+++ core/src/state/map_state.rs
@@ -13,70 +13,26 @@ pub struct MapWriter<W, WM> {
   pub(super) part_map: WM,
 }
 
-pub struct MapWriterAsReader<W, M> {
-  pub(super) origin: W,
-  pub(super) part_map: M,
-}
-
-impl<S, V: ?Sized, M> StateReader for MapReader<S, M>
+impl<S, M, V: ?Sized> StateReader for MapReader<S, M>
 where
   Self: 'static,
   S: StateReader,
-  M: Fn(&S::Value) -> PartRef<V> + Clone,
+  M: MapReaderFn<S::Value, Output = V>,
 {
   type Value = V;
-  type OriginReader = S;
   type Reader = MapReader<S::Reader, M>;
 
   #[inline]
-  fn read(&self) -> ReadRef<Self::Value> { ReadRef::map(self.origin.read(), &self.part_map) }
+  fn read(&self) -> ReadRef<Self::Value> { self.part_map.call(self.origin.read()) }
+
+  #[inline]
+  fn clone_boxed_reader(&self) -> Box<dyn StateReader<Value = Self::Value>> {
+    Box::new(self.clone_reader())
+  }
 
   #[inline]
   fn clone_reader(&self) -> Self::Reader {
     MapReader { origin: self.origin.clone_reader(), part_map: self.part_map.clone() }
-  }
-
-  #[inline]
-  fn origin_reader(&self) -> &Self::OriginReader { &self.origin }
-
-  #[inline]
-  fn try_into_value(self) -> Result<Self::Value, Self>
-  where
-    Self::Value: Sized,
-  {
-    Err(self)
-  }
-}
-
-impl<S, V: ?Sized, M> StateReader for MapWriterAsReader<S, M>
-where
-  Self: 'static,
-  S: StateReader,
-  M: Fn(&mut S::Value) -> PartMut<V> + Clone,
-{
-  type Value = V;
-  type OriginReader = S;
-  type Reader = MapWriterAsReader<S::Reader, M>;
-
-  #[inline]
-  fn read(&self) -> ReadRef<Self::Value> {
-    ReadRef::mut_as_ref_map(self.origin.read(), &self.part_map)
-  }
-
-  #[inline]
-  fn clone_reader(&self) -> Self::Reader {
-    MapWriterAsReader { origin: self.origin.clone_reader(), part_map: self.part_map.clone() }
-  }
-
-  #[inline]
-  fn origin_reader(&self) -> &Self::OriginReader { &self.origin }
-
-  #[inline]
-  fn try_into_value(self) -> Result<Self::Value, Self>
-  where
-    Self::Value: Sized,
-  {
-    Err(self)
   }
 }
 
@@ -87,8 +43,7 @@ where
   M: Fn(&mut S::Value) -> PartMut<V> + Clone,
 {
   type Value = V;
-  type OriginReader = S;
-  type Reader = MapWriterAsReader<S::Reader, M>;
+  type Reader = MapReader<S::Reader, WriterMapReaderFn<M>>;
 
   #[inline]
   fn read(&self) -> ReadRef<Self::Value> {
@@ -96,19 +51,16 @@ where
   }
 
   #[inline]
-  fn clone_reader(&self) -> Self::Reader {
-    MapWriterAsReader { origin: self.origin.clone_reader(), part_map: self.part_map.clone() }
+  fn clone_boxed_reader(&self) -> Box<dyn StateReader<Value = Self::Value>> {
+    Box::new(self.clone_reader())
   }
 
   #[inline]
-  fn origin_reader(&self) -> &Self::OriginReader { &self.origin }
-
-  #[inline]
-  fn try_into_value(self) -> Result<Self::Value, Self>
-  where
-    Self::Value: Sized,
-  {
-    Err(self)
+  fn clone_reader(&self) -> Self::Reader {
+    MapReader {
+      origin: self.origin.clone_reader(),
+      part_map: WriterMapReaderFn(self.part_map.clone()),
+    }
   }
 }
 
@@ -118,9 +70,20 @@ where
   W: StateWriter,
   M: Fn(&mut W::Value) -> PartMut<V> + Clone,
 {
+  type Watcher = Watcher<Self::Reader>;
+  #[inline]
+  fn clone_boxed_watcher(&self) -> Box<dyn StateWatcher<Value = Self::Value>> {
+    Box::new(self.clone_watcher())
+  }
+
   #[inline]
   fn raw_modifies(&self) -> CloneableBoxOp<'static, ModifyScope, Infallible> {
     self.origin.raw_modifies()
+  }
+
+  #[inline]
+  fn clone_watcher(&self) -> Watcher<Self::Reader> {
+    Watcher::new(self.clone_reader(), self.raw_modifies())
   }
 }
 
@@ -130,13 +93,10 @@ where
   W: StateWriter,
   M: Fn(&mut W::Value) -> PartMut<V> + Clone,
 {
-  type Writer = MapWriter<W::Writer, M>;
-  type OriginWriter = W;
-
   fn into_reader(self) -> Result<Self::Reader, Self> {
     let Self { origin, part_map } = self;
     match origin.into_reader() {
-      Ok(origin) => Ok(MapWriterAsReader { origin, part_map }),
+      Ok(origin) => Ok(MapReader { origin, part_map: WriterMapReaderFn(part_map) }),
       Err(origin) => Err(Self { origin, part_map }),
     }
   }
@@ -153,12 +113,14 @@ where
   }
 
   #[inline]
-  fn clone_writer(&self) -> Self::Writer {
-    MapWriter { origin: self.origin.clone_writer(), part_map: self.part_map.clone() }
+  fn clone_boxed_writer(&self) -> Box<dyn StateWriter<Value = Self::Value>> {
+    Box::new(self.clone_writer())
   }
 
   #[inline]
-  fn origin_writer(&self) -> &Self::OriginWriter { &self.origin }
+  fn clone_writer(&self) -> Self {
+    MapWriter { origin: self.origin.clone_writer(), part_map: self.part_map.clone() }
+  }
 }
 
 impl<V: ?Sized, S, F> RenderProxy for MapReader<S, F>
@@ -169,16 +131,6 @@ where
   V: Render,
 {
   #[inline]
-  fn proxy(&self) -> impl Deref<Target = impl Render + ?Sized> { self.read() }
-}
-
-impl<V: ?Sized, S, F> RenderProxy for MapWriterAsReader<S, F>
-where
-  Self: 'static,
-  S: StateReader,
-  F: Fn(&mut S::Value) -> PartMut<V> + Clone,
-  V: Render,
-{
   fn proxy(&self) -> impl Deref<Target = impl Render + ?Sized> { self.read() }
 }
 
@@ -196,3 +148,31 @@ where
 {
   fn into_widget_strict(self) -> Widget<'static> { Compose::compose(self) }
 }
+
+trait MapReaderFn<Input: ?Sized>: Clone {
+  type Output: ?Sized;
+  fn call<'a>(&self, input: ReadRef<'a, Input>) -> ReadRef<'a, Self::Output>;
+}
+
+impl<Input: ?Sized, Output: ?Sized, F> MapReaderFn<Input> for F
+where
+  F: Fn(&Input) -> PartRef<Output> + Clone,
+{
+  type Output = Output;
+  fn call<'a>(&self, input: ReadRef<'a, Input>) -> ReadRef<'a, Self::Output> {
+    ReadRef::map(input, self)
+  }
+}
+
+impl<Input: ?Sized, Output: ?Sized, F> MapReaderFn<Input> for WriterMapReaderFn<F>
+where
+  F: Fn(&mut Input) -> PartMut<Output> + Clone,
+{
+  type Output = Output;
+  fn call<'a>(&self, input: ReadRef<'a, Input>) -> ReadRef<'a, Self::Output> {
+    ReadRef::mut_as_ref_map(input, &self.0)
+  }
+}
+
+#[derive(Clone)]
+pub struct WriterMapReaderFn<F>(pub(crate) F);