+++ core/src/state/splitted_state.rs
@@ -19,8 +19,7 @@ where
   W: Fn(&mut O::Value) -> PartMut<V> + Clone,
 {
   type Value = V;
-  type OriginReader = O;
-  type Reader = MapWriterAsReader<O::Reader, W>;
+  type Reader = MapReader<O::Reader, WriterMapReaderFn<W>>;
 
   #[track_caller]
   fn read(&self) -> ReadRef<Self::Value> {
@@ -28,19 +27,16 @@ where
   }
 
   #[inline]
-  fn clone_reader(&self) -> Self::Reader {
-    MapWriterAsReader { origin: self.origin.clone_reader(), part_map: self.splitter.clone() }
+  fn clone_boxed_reader(&self) -> Box<dyn StateReader<Value = Self::Value>> {
+    Box::new(self.clone_reader())
   }
 
   #[inline]
-  fn origin_reader(&self) -> &Self::OriginReader { &self.origin }
-
-  #[inline]
-  fn try_into_value(self) -> Result<Self::Value, Self>
-  where
-    Self::Value: Sized,
-  {
-    Err(self)
+  fn clone_reader(&self) -> Self::Reader {
+    MapReader {
+      origin: self.origin.clone_reader(),
+      part_map: WriterMapReaderFn(self.splitter.clone()),
+    }
   }
 }
 
@@ -50,8 +46,19 @@ where
   O: StateWriter,
   W: Fn(&mut O::Value) -> PartMut<V> + Clone,
 {
+  type Watcher = Watcher<Self::Reader>;
+
+  #[inline]
+  fn clone_boxed_watcher(&self) -> Box<dyn StateWatcher<Value = Self::Value>> {
+    Box::new(self.clone_watcher())
+  }
   fn raw_modifies(&self) -> CloneableBoxOp<'static, ModifyScope, std::convert::Infallible> {
     self.info.notifier.raw_modifies().box_it()
+  }
+
+  #[inline]
+  fn clone_watcher(&self) -> Watcher<Self::Reader> {
+    Watcher::new(self.clone_reader(), self.raw_modifies())
   }
 }
 
@@ -61,9 +68,6 @@ where
   O: StateWriter,
   W: Fn(&mut O::Value) -> PartMut<V> + Clone,
 {
-  type Writer = SplittedWriter<O::Writer, W>;
-  type OriginWriter = O;
-
   fn into_reader(self) -> Result<Self::Reader, Self> {
     if self.info.writer_count.get() == 1 { Ok(self.clone_reader()) } else { Err(self) }
   }
@@ -77,7 +81,12 @@ where
   #[inline]
   fn shallow(&self) -> WriteRef<Self::Value> { self.split_ref(self.origin.shallow()) }
 
-  fn clone_writer(&self) -> Self::Writer {
+  #[inline]
+  fn clone_boxed_writer(&self) -> Box<dyn StateWriter<Value = Self::Value>> {
+    Box::new(self.clone_writer())
+  }
+
+  fn clone_writer(&self) -> Self {
     self.info.inc_writer();
     SplittedWriter {
       origin: self.origin.clone_writer(),
@@ -85,9 +94,6 @@ where
       info: self.info.clone(),
     }
   }
-
-  #[inline]
-  fn origin_writer(&self) -> &Self::OriginWriter { &self.origin }
 }
 
 impl<V: ?Sized, O, W> SplittedWriter<O, W>