+++ core/src/state/watcher.rs
@@ -24,16 +24,17 @@ impl<R> From<Watcher<Reader<R>>> for Reader<R> {
 impl<R: StateReader> StateReader for Watcher<R> {
   type Value = R::Value;
   type Reader = R::Reader;
-  type OriginReader = R::OriginReader;
 
   #[inline]
   fn read(&self) -> ReadRef<Self::Value> { self.reader.read() }
 
   #[inline]
-  fn clone_reader(&self) -> Self::Reader { self.reader.clone_reader() }
+  fn clone_boxed_reader(&self) -> Box<dyn StateReader<Value = Self::Value>> {
+    Box::new(self.clone_reader())
+  }
 
   #[inline]
-  fn origin_reader(&self) -> &Self::OriginReader { self.reader.origin_reader() }
+  fn clone_reader(&self) -> Self::Reader { self.reader.clone_reader() }
 
   #[inline]
   fn try_into_value(self) -> Result<Self::Value, Self>
@@ -48,8 +49,19 @@ impl<R: StateReader> StateReader for Watcher<R> {
 }
 
 impl<R: StateReader> StateWatcher for Watcher<R> {
+  type Watcher = Watcher<R::Reader>;
+
+  #[inline]
+  fn clone_boxed_watcher(&self) -> Box<dyn StateWatcher<Value = Self::Value>> {
+    Box::new(self.clone_watcher())
+  }
+
   #[inline]
   fn raw_modifies(&self) -> CloneableBoxOp<'static, ModifyScope, Infallible> {
     self.modifies_observable.clone()
+  }
+
+  fn clone_watcher(&self) -> Watcher<Self::Reader> {
+    Watcher::new(self.clone_reader(), self.raw_modifies())
   }
 }